looses precision fix

+++ src/FPDF_Protection.php
@@ -141,7 +141,7 @@ class FPDF_Protection extends tFPDF
         protected function _RC4($key, $text)
         {
                 if ($this->last_rc4_key != $key) {
-                        $k = str_repeat($key, 256/strlen($key)+1);
+                        $k = str_repeat($key, (int)(256/strlen($key)+1));
                         $rc4 = range(0,255);
                         $j = 0;
                         for ($i=0; $i<256; $i++){
